requirements.txt
langchain-text-splitters
PyPDF2
python-docx
pandas
chromadb
ollama
xlwings
